# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel CRUD Users & Comments API",
        "description": "Документация к API аутентификации, постов и комментариев",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить список постов",
                "operationId": "cce7863441436fd49dc474491f658292",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Статус поста",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ],
                            "example": "active"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID автора поста",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество записей на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список постов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Создать пост",
                "operationId": "21a6ac5f9280a3cf958c3d066b5600ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пост успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить пост по ID",
                "operationId": "6b9af649206ee145dfc5584522c37327",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о посте",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Обновить пост",
                "operationId": "56b92780703545f9cdee51b2f0bfd2f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пост успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "403": {
                        "description": "Нет прав для обновления поста"
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Удалить пост",
                "operationId": "a879f12c52aaa72acee21585bf617046",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Пост успешно удален"
                    },
                    "403": {
                        "description": "Нет прав для удаления поста"
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{userId}/active": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить активные посты пользователя",
                "operationId": "42c7cdc6285bb25a60e70bbfab09effe",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество записей на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список активных постов пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts/mine": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить мои посты",
                "operationId": "30d9c4a38821f7261f0c26493635cb79",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество записей на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список моих постов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PostStoreRequest": {
                "required": [
                    "body"
                ],
                "properties": {
                    "body": {
                        "type": "string",
                        "example": "Содержимое нового поста"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    }
                },
                "type": "object"
            },
            "PostUpdateRequest": {
                "properties": {
                    "body": {
                        "type": "string",
                        "example": "Обновленное содержимое поста"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "body": {
                        "type": "string",
                        "example": "Содержимое поста"
                    },
                    "author": {
                        "$ref": "#/components/schemas/User"
                    },
                    "comments_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Введите Bearer токен",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}